requirements.txt
Django>=5.2
djangorestframework>=3.14
drf-spectacular>=0.27
whitenoise>=6.6
gunicorn>=21.0
